+++ gameserver/src/main/java/com/lhf/game/creature/intelligence/actionChoosers/VocationChooser.java
@@ -1,4 +1,4 @@
-package com.lhf.game.creature.intelligence.targetChoosers;
+package com.lhf.game.creature.intelligence.actionChoosers;
 
 import java.util.List;
 import java.util.Map;
